fast_api/requirements.txt
fastapi
pymongo
uvicorn
passlib
PyJWT
pydantic
typing
pytz